Breaking Up the Monolitic Engine Tests
That was a lot of code. Phew.
